+++ NEWS.md
@@ -1,4 +1,11 @@
 # scMerge 0.99 (development version)
+
+## scMerge 0.99.19
+* Fixed spelling
+* Added verbose option
+* Code coverage at 75 percent (more tests on error handling needed)
+
+
 ## scMerge 0.99.17 
 * Fixed README `install_github` vignette issue. 
 * Fixed pkgdown organisation issue.